Move expression classes to a separate package
