+++ src/main/kotlin/sigma/DictAssociativeTable.kt
@@ -1,5 +1,7 @@
 package sigma
 
+import sigma.expressions.Expression
+
 class DictAssociativeTable(
     override val environment: Table,
     associations: Map<Value, Expression>,
