+++ src/main/kotlin/sigma/IntValue.kt
@@ -1,7 +1,5 @@
 package sigma
 
-import sigma.parser.antlr.SigmaParser
-
 data class IntValue(
     val value: Int,
 ) : PrimitiveValue() {
@@ -195,26 +193,4 @@ data class IntValue(
     }
 
     override fun dump(): String = value.toString()
-}
-
-data class IntLiteral(
-    val value: IntValue,
-) : Expression {
-    companion object {
-        fun of(value: Int) = IntLiteral(
-            value = IntValue(value)
-        )
-
-        fun build(
-            ctx: SigmaParser.IdentifierContext,
-        ): SymbolLiteral = SymbolLiteral(
-            symbol = Symbol(name = ctx.text),
-        )
-    }
-
-    override fun evaluate(
-        context: Table,
-    ): Value = value
-
-    override fun dump(): String = value.dump()
 }
