+++ src/main/kotlin/sigma/Value.kt
@@ -1,5 +1,7 @@
 package sigma
 
+import sigma.expressions.Expression
+
 // Idea: Hierarchy like Expression -> Thunk -> Value?
 sealed class Value : Thunk {
     final override fun obtain(): Value = this
