+++ src/main/kotlin/sigma/expressions/Expression.kt
@@ -1,8 +1,13 @@
-package sigma
+package sigma.expressions
 
 import org.antlr.v4.runtime.CharStreams
 import org.antlr.v4.runtime.CommonTokenStream
 import org.antlr.v4.runtime.ParserRuleContext
+import sigma.GlobalContext
+import sigma.Symbol
+import sigma.Table
+import sigma.Thunk
+import sigma.Value
 import sigma.parser.antlr.SigmaLexer
 import sigma.parser.antlr.SigmaParser
 import sigma.parser.antlr.SigmaParser.AbstractionAltContext
