+++ src/main/kotlin/sigma/expressions/TypeReference.kt
@@ -1,4 +1,8 @@
-package sigma
+package sigma.expressions
+
+import sigma.Symbol
+import sigma.Type
+import sigma.TypeContext
 
 data class TypeReference(
     val referee: Symbol,
