+++ src/test/kotlin/sigma/ExpressionEvaluationTests.kt
@@ -1,5 +1,6 @@
 package sigma
 
+import sigma.expressions.Expression
 import kotlin.test.Test
 import kotlin.test.assertEquals
